pyproject.toml
[project]
name = "appagent-claw"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "mss>=10.1.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pyautogui>=0.9.54",
    "pynput>=1.8.1",
    "pytest>=9.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
]
